int to float

+++ worker.py
@@ -4,7 +4,7 @@ from threading import Event
 resonate = Resonate.remote(group="worker")
 
 @resonate.register
-def sleeping_workflow(ctx: Context, wf_id: str, secs: int):
+def sleeping_workflow(ctx: Context, wf_id: str, secs: float):
     print(f"Workflow {wf_id} starting, will sleep for {secs} seconds.")
     yield ctx.sleep(secs)
     return f"Workflow {wf_id} completed after sleeping for {secs} seconds."
